+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Remove private methods from cube
 - Remove deprecated methods
 - Fixed FactTable spec
-- Added SubCube with back link to original cube `originalCube` for chaining between operations
+- Added SubCube as a result of operations with back links `originalCube` and `previousCube` for chaining between operations
 
 *0.15.0*
 - Removed `getCellsBySet`, `getFactsBySet` and `getDimensionMembersBySet`